Add precompiled tags if backedge is missing
In https://github.com/JuliaLang/julia/issues/49617 it was found that
not all callees have backedges to callers. There is some speculation
that rather than adding the backedge to those cases, we may in fact
be able to prune the backedge graph more substantially in cases where
it may be safe to avoid propagating invalidation.

Consequently, to enforce the guarantees made by `@compile_workload`,
PrecompileTools should check the inference graph and add tags in cases
where a backedge is missing.