Ensure compilation happens after logging starts
If compilation of the workload happens before the logging system is
initialized, many CodeInstances may be missed. This exploits
`Core.@invokelatest` to prevent inference from proceeding into the
workload. It removes the `@force_compile`, which introduces
some risk that even the workload will be interpreted.

Also wraps the `@setup_workload` block with a `let`.

Fixes #17
Fixes #26
Fixes #37